python_api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
espn-api==0.45.1
requests
urllib3<=2.2.3